geometry:optimization: Split ConvexSet into multiple files
There is zero code change here.  It only splits the existing code into multiple files (as suggested by @jwnimmer-tri)
