+++ geometry/optimization/hpolyhedron.h
@@ -1,0 +1,112 @@
+#pragma once
+
+#include <memory>
+#include <optional>
+#include <utility>
+
+#include "drake/geometry/optimization/convex_set.h"
+#include "drake/geometry/optimization/hyperellipsoid.h"
+
+namespace drake {
+namespace geometry {
+namespace optimization {
+
+/** Implements a polyhedral convex set using the half-space representation:
+`{x| A x ≤ b}`.  Note: This set may be unbounded.
+@ingroup geometry_optimization
+*/
+class HPolyhedron final : public ConvexSet {
+ public:
+  DRAKE_DEFAULT_COPY_AND_MOVE_AND_ASSIGN(HPolyhedron)
+
+  /** Constructs the polyhedron.
+  @pre A.rows() == b.size().
+  */
+  HPolyhedron(const Eigen::Ref<const Eigen::MatrixXd>& A,
+              const Eigen::Ref<const Eigen::VectorXd>& b);
+
+  /** Constructs a new HPolyhedron from a SceneGraph geometry and pose in the
+  @p expressed_in frame, obtained via the QueryObject.  If @p expressed_in
+  frame is std::nullopt, then it will be expressed in the world frame.
+
+  @throws std::exception the geometry is not a convex polytope. */
+  HPolyhedron(const QueryObject<double>& query_object, GeometryId geometry_id,
+              std::optional<FrameId> expressed_in = std::nullopt);
+  // TODO(russt): Add a method/constructor that would create the geometry using
+  // SceneGraph's AABB or OBB representation (for arbitrary objects) pending
+  // #15121.
+
+  virtual ~HPolyhedron() {}
+
+  /** Returns the half-space representation matrix A. */
+  const Eigen::MatrixXd& A() const { return A_; }
+
+  /** Returns the half-space representation vector b. */
+  const Eigen::VectorXd& b() const { return b_; }
+
+  // TODO(russt): Add bool IsBounded() so users can test the precondition.
+  /** Solves a semi-definite program to compute the inscribed ellipsoid.
+  From Section 8.4.2 in Boyd and Vandenberghe, 2004, we solve
+  @verbatim
+  max_{C,d} log det (C)
+        s.t. |aᵢC|₂ ≤ bᵢ - aᵢd, ∀i
+            C ≽ 0
+  @endverbatim
+  where aᵢ and bᵢ denote the ith row.  This defines the ellipsoid
+  E = { Cx + d | |x|₂ ≤ 1}.
+
+  @pre the HPolyhedron is bounded.
+  @throws std::exception if the solver fails to solve the problem.
+  */
+  HyperEllipsoid MaximumVolumeInscribedEllipsoid() const;
+
+  /** Constructs a polyhedron as an axis-aligned box from the lower and upper
+   * corners. */
+  static HPolyhedron MakeBox(const Eigen::Ref<const Eigen::VectorXd>& lb,
+                             const Eigen::Ref<const Eigen::VectorXd>& ub);
+
+  /** Constructs the L∞-norm unit box in @p dim dimensions, {x | |x|∞ <= 1 }.
+  This is an axis-aligned box, centered at the origin, with edge length 2. */
+  static HPolyhedron MakeUnitBox(int dim);
+
+ private:
+  bool DoPointInSet(const Eigen::Ref<const Eigen::VectorXd>& x,
+                    double tol) const final;
+
+  void DoAddPointInSetConstraint(
+      solvers::MathematicalProgram* prog,
+      const Eigen::Ref<const solvers::VectorXDecisionVariable>& vars)
+      const final;
+
+  // TODO(russt): Implement DoToShapeWithPose.  Currently we don't have a Shape
+  // that can consume this output.  The obvious candidate is Convex, that class
+  // currently only stores the filename of an .obj file, and I shouldn't need
+  // to go to disk and back to get this geometry into ProximityEngine nor
+  // MeshcatVisualizer.
+  //
+  // When we do implement this and also have the AHPolyhedron class, we should
+  // add recommendation here that: "If ambient_dimension() != 3, then consider
+  // using the AH polytope representation to project it to 3D."
+  std::pair<std::unique_ptr<Shape>, math::RigidTransformd> DoToShapeWithPose()
+      const final {
+    throw std::runtime_error(
+        "ToShapeWithPose is not implemented yet for HPolyhedron.  Implementing "
+        "this will likely require additional support from the Convex shape "
+        "class (to support in-memory mesh data, or file I/O).");
+  }
+
+  // Implement support shapes for the ShapeReifier interface.
+  using ShapeReifier::ImplementGeometry;
+  void ImplementGeometry(const HalfSpace&, void* data) final;
+  void ImplementGeometry(const Box& box, void* data) final;
+  // TODO(russt): Support ImplementGeometry(const Convex& convex, ...), but
+  // currently it would require e.g. digging ReadObjForConvex out of
+  // proximity_engine.cc.
+
+  Eigen::MatrixXd A_{};
+  Eigen::VectorXd b_{};
+};
+
+}  // namespace optimization
+}  // namespace geometry
+}  // namespace drake
